Add uart_client as option.
It will use minicom to receive UART input from bluenet devices. For now
it is used that /dev/ACM0 exists. This has to be made configurable.

The installation of minicom is not enforced. This depends on the
distribution.

The INSTALL.md document is updated to signal that this option is
available.