Set temperature thresholds for the NTC near the dimmer.
The NTC is placed at the other side of the PCB right opposite
of the diodes that in reverse with the IGBTs.

The current values lead to a temperature warning. This is something
that has to be investigated.

An NTC has a resistance that lowers with higher temperatures. It is
used in a voltage divider with 15k towards 3.3V. If the resistance gets
lower, the voltage drop over the NTC becomes lower and V_temp becomes
lower.

The NTC used is the Murata NCP15XH103J03RC. It is at 10k on room
temperature. V_temp is then at 10/(10+15)*3V3 = 1.32V. With lower
resistance than 10k we will measure lower voltages. Hence, temperature
and voltage are "inverted". This is also the configuration in this
commit.

Needs to be checked!!