Updates to pin layout for ACR0B11A board.
+ Not all the pins use the same numbering. All AIN pins were incorrect.
+ Adjusted temp thresholds slightly. Now the temp uses the right pin and
it works. Fixes
https://github.com/crownstone/bluenet/commit/fc3a451471f911a648d7e715ab476767b6796fa2.
+ First attempt to set voltage range and multiplier. This needs to be
checked further in-circuit.

TODO: Use the same numbering for all the pins and map in the ADC file to
AIN pin numbers. Also check there if it is an AIN pin. If we use the
same numbering scheme we can quickly see if we have accidentally used
the same pin twice.