Working version, bootloader version is retrieved.
The bootloader version is communicated to the application using a
dedicated RAM section.

+ Separate RAM memory block for data exchange.
+ Made the struct packed to reduce conflicts.
+ Made the first index equal to 1 to reduce change to find default 0
index while nothing is there.
+ Not yet implemented checksum, but is trivial to add.
+ The bootloader version is only printed to serial, it's not yet written
to state (or flash with FDS).

This completes the proof of concept implementation for
https://github.com/crownstone/bluenet/issues/91. The roundtrip to the
bootloader can be prevented with exactly this method.

TODO: Observed that only flashing the new application seem to trip up
the bootloader. Have to check if that has to do with the changes in this
branch or if this already happens anyway. If it has to do with this
branch it might be a checksum over RAM(?), but that would be weird.