Working on IPC from bootloader to app
+ Create an array in the RAM section, not just a single struct.
+ Implement a checksum function. It includes index. It is a sum. It
inverts the total.
+ Have some predefined macros about size and # of structs.
+ Lift RAM_BOOTLOADER_LENGTH to general config file.
+ Add explanation to crownstone.defs.cmake to which macros are needed (a
lot of them are not).
+ Removed most macros having to do with memory management (e.g.
RAM_BOOTLOADER_LENGTH is read from the default config and does not need
to be set as macro, nor added to the cache).

TODO: There are a lot of definitions that can be removed. If necessary
from the code base they can be added to cs_Autoconfig. Removing most
set() statements will lead to less conflicts with the cache.

TODO: The application does not yet use the array in the RAM section, not
checks the checksum. (My USB port stopped working. Probably because my
adapter is failing. Ordered new adapter. Never buy a Lenovo laptop.)