Bootloader and app, make dfu work again.
This concerns error

```
<warning> nrf_sdh_ble: Insufficient RAM allocated for the SoftDevice.
<warning> nrf_sdh_ble: Change the RAM start location from 0x20002380 to 0x20002A98.
<warning> nrf_sdh_ble: Maximum RAM size for application is 0x1D568.
<error> nrf_sdh_ble: sd_ble_enable() returned NRF_ERROR_NO_MEM.
```

+ The bootloader requires more RAM than the application.
+ Introduce RAM_BOOTLOADER_EXPANSION variable to indicate this.
+ Also have location of IPC part of RAM at the end of RAM rather than
the start.

The latter is according to the recommendations at
https://devzone.nordicsemi.com/f/nordic-q-a/33530/ram-variables-exchange-between-bootloader-and-application.
The stack grows from the top down and this does not run as much risk of
being overwritten by the stack. I think this is too far-fetched because
what about the heap, but anyway. We also do not require protection

= Protection

Also tried to protect more of the flash to get rid of error:

```
<error> app: Could not protect SoftDevice and application, 0x7.
```

See https://devzone.nordicsemi.com/f/nordic-q-a/53020/sdk-15-3-0-non-fatal-protection-error-during-nrf_bootloader_app_start_final

+ We introduce ALIGN_TO_PAGE
+ Note we try to protect from 0 to end of bootloader.

TODO: Is this too much? In the code MBR and bootloader is separately
protected. Do we on purpose not protect in-between for "intermediate bootloaders"?