Microapp is received over BLE
+ The app gets written to flash.
+ The settings are such that the transmission unit is large enough.
+ First the characteristic returns ERR_BUSY, after fstorage is done
ERR_SUCCESS is written.
+ The listener might not receive every notification. Implement a repeat
mechanism, so a notifiation is sent at least N times (currently 3).
+ The app_checksum is deleted. This makes the data word-aligned. It also
reduces the header space (MTU very large takes up RAM, also in mesh
notifications).
+ The app is not written correctly yet. The checksum is not checked yet.

However, for future work, when printing the data on the moment that it
is received, it is still correct. The sender likely already sends a new
message when the previous has not been received.

How to solve: add a challenge to the response. If the next chunk does
not include the challenge of the previous reponse, we do not accept it.
The "acceptance" should be at the moment the buffer is written. Can that
actually be prevented? If correct, there is a copy action though, so we
can make sure that is protected against multiple successive rights.