Split microapp code in several files.
Create microapp, microapp protocol, and microapp storage files.

+ cs_MicroApp is the general file, gets ticks from scheduler and implements the OTA protocol
+ cs_MicroAppProtocol concerns the protocol between the bluenet code and the microapp
+ cs_MicroAppStorage is about storing the microapp to flash

There is some back and forth between storage and protocol. For example,
currently the storage class **enables** the microapp, while the protocol
class governs the **runtime** information. If a microapp is enabled has
to be stored in flash (or it will be started the next way the bluenet
code boots).

TODO: There is still some love required around booting and checking.
Currently the microapp will always be **disabled** at boot.