Introduce phony target make write_config
This will write the information to e.g. set up a Crownstone to the file
CMakeBuild.dynamic.config locally.

On running `make setup` a json file is generated with this information
(as well). For example the MAC address.

You can also only write particular fields with e.g.:

	BLUENET_CONFIG=MAC_ADDRESS make write_config

This will read the MAC address and then write it to the file (if it is
different).