Implement callbacks by writing to ipc buffer
Previously callbacks have been implemented by registering callbacks in
the form of function pointers at bluenet. Then bluenet was actually
calling these methods. This happened by accident in the coroutine
context of bluenet itself while it should have happened within the
coroutine context of the microapp.

This uses the interproces buffer to write the data to and then calls the
microapp upon which a proper context switch takes place.

For future reference, app size is 263780 (0x40664) with meshing enabled,
with microapps enabled, with non-binary protocol, and with warning level
logs.

This implements callbacks for gpios and incoming mesh messages.

The microapp code has to be updated in tandem.

This commit also introduces a couple of new structs so now manually
payload[x] array inspection is required. Just cast the buffer and read
or write the relevant fields.

For callback messages, it is assumed that the first part is according to
`microapp_cmd_acked_t`, that is, a `cmd` field and an `ack` field.