Rewrite handling mesh messages for microapp
In particular a rewrite of handleMicroappMeshCommand so that it uses
complete structs such as microapp_mesh_read_cmd_t which simplifies the
code.

Most importantly some comments are placed about the hidden assumption
that this is a synchronous call which is not the case. The data written
to the buffer is only handled by the microapp after calling
callMicroapp(). If not immediately called, the buffer might be
overwritten by other functions in cs_MicroappProtocol.cpp.

There are (at least) two ways to fix this. One way is to call
callMicroapp() immediately, making it a semi-synchronous sequence. The
second way is to create a ipc buffer with multiple entries.

The first option means that after a microapp call to available() there
will be a context switch to bluenet and then back to the microapp and
then back to bluenet. The second option means that we use more memory
for inter-process communication.

We have limited RAM, hence the recommendation is to for option one.

This is not yet implemented.

---
App size 263764 (0x40654)