Use microapp cmd structs rather than raw buffers
On many spots the payload from the microapp to the bluenet code was
directly read. For example the opcode was read from payload[0] etc.
This is not easy to maintain:

+ A command cannot easily add a field without having to byte-shift
everything on both the sending and receiving side.
+ If structs are used and a protocol version gets bumped to indicate a
breaking change, it is likely only a matter of recompilation (no changes
in indexes into a raw buffer have to be updated/maintained).
+ A struct can be easily checked on size limitations. No need for a
separate size field (which is under control of the microapp anyway).
+ Casting to the right struct and then getting the right field is much
easier for a developer and bugs are found faster.
+ It is easier to reuse parts of structs for similar messages (e.g. the
log messages).

Some minor things like introducing maximum number of profiles and
preventing buffer overflow initiated from the microapp.

Added some comments on the use of printf (or logs of floating point
values) in plain-text mode. This is not available considering size
limitations. Note, it does crash the firmware if attempted!

This is tested with the basic.ino microapp example code. Functionality
like the logs, delays, etc. are all operational.

Callbacks into the microapp (e.g. forwarding incoming mesh messages)
must still be tested.

-- The app is of size 209124 (0x330E4)