Fully functional GPIO microapp interrupts
The interrupts for GPIO are completely functional. They can be
configured from the microapp, will be propagated through bluenet and end
up at the GPIO driver. A callback from bluenet is set into IPC RAM and
then a context switch takes place. The microapp acknowledges the
callback, and then resumes execution, starting with the callback itself.
If there is another interrupt this will first end up in bluenet and will
only end up at the microapp at a later stage.

It is possible to write Serial.println("Hello") statements in the
interrupt service routine. It is not optimized for latency, but it is
perceptually instantaneous from pressing a button on the development kit
to seeing "Toggle" printed on the command line. See here for example an
interrupt using a `CHANGE` event (button press and release both generate
an event):

```
[2022-01-31 18:17:51.487] [source/src/drivers/cs_Gpio.cpp : 228 ] GPIO event on pin 11
[2022-01-31 18:17:51.531] [croapp/cs_MicroappProtocol.cpp : 482 ] GPIO callback for microapp pin 4
[2022-01-31 18:17:51.533] [croapp/cs_MicroappProtocol.cpp : 701 ] Toggle

[2022-01-31 18:17:51.552] [source/src/drivers/cs_Gpio.cpp : 228 ] GPIO event on pin 11
[2022-01-31 18:17:51.635] [croapp/cs_MicroappProtocol.cpp : 482 ] GPIO callback for microapp pin 4
[2022-01-31 18:17:51.638] [croapp/cs_MicroappProtocol.cpp : 701 ] Toggle
```

There's roughly a 100 ms delay. The code in cs_MicroappProtocol.cpp gets
events using the app scheduler and there's where this delay comes from.
For now there is no reason to optimize this.

The code is run on the PCA10056 (nRF52840) because there's more space
for microapps on it. Corresponding cs_Boards.c file is not committed
because there is a pull request to be accepted: https://github.com/crownstone/bluenet/pull/149

The code is compatible with the crownstone-microapp repository commit: https://github.com/crownstone/crownstone-microapp/commit/114fe533de73d337e9ea8c539e65e008e53f21c1

-- The app is of size 216752 (0x34EB0)