Implement Bluetooth functionality for microapps
+ Halfway supporting Bluetooth advertisements. This is on the nRF52840
which requires Bluetooth Mesh 5.0.0 which still has some bugs that are
solved in the meantime. Some of it has to do with actually implementing
persistence of data. Some of it has to do with additional filtering
within the mesh 5.0.0 code compared to the 3.2.0 code.
+ Adding code for connections. Not tested yet.

This commit is perhaps premature, but it would benefit from being merged
with https://github.com/crownstone/bluenet/tree/mesh-sdk-5, hence this
is done first.