Restructuring microapp setup and cmake
The coroutine code sets aside a piece of memory before the coroutine is
started to store stack information of bluenet to be able to jump back.
Until now we have communicated this information back and forth over IPC
and subsequenly as arguments to e.g. the yield function. It is much
cleaner to just store this information in cs_DoubleStackCoroutine though
for which a setCoroutineContext call is created. This cleans up a lot of
code.

In the meantime also cmake is slightly adjusted to start using templates
to fix https://github.com/crownstone/bluenet/issues/155.