Microapps wrap-up
There is now a separate counter for successful and non-successful
callback reception by the microapp. There is a buffer at the microapp
that can be used to be able to handle more than a single callback
simultaneously. This has the disadvantage that every message will be
memcopied (if this buffer is not full). However, that seems a small
price to pay for the ability to keep sending messages towards a microapp
that is fast in processing them.

Review:

+ Documentation can benefit from some love.
+ The code has come a long way. Other functions might have suffered. For
example GPIO events have to be checked again if they still work.
+ There is still optimization and clean up possible, but this
implementation is a true improvement w.r.t. the previous implementation
which did start a new coroutine for each loop and didn't use one for
soft interrupt service routines at all.
+ The code has been extensively run on the nRF52840 and might need some
attention w.r.t. the nRF52832.

It would be nice to have a swift review process. If there are major
rewrites required I would propose to do that in another PR.