Introduce BUILD_P2P_DFU define
The default is off. If enabled, both the bootloader and the firmware
will obtain peer to peer DFU (device firmware upgrade) functionality.

Workflow:

1. Have one ordinary Crownstone (this is the DFU target).
2. Have one with P2P DFU functionality.

Crownstone 2 will use (for now) a hardcoded MAC address to update the
Crownstone 1. After a timeout Crownstone 2 will connect with Crownstone
1 and start the DFU protocol. For this it requires a so-called "init
packet". Crownstone 2 can therefore best be flashed via a DFU tool
rather than via nrfjprog. The bootloader is adjusted so that this init
packet is stored in flash and subsequently available to the bluenet
application.

There is a python script to upload an application which can be found at
https://github.com/crownstone/crownstone-lib-python-ble/blob/ble-dfu-tool-179086349/tools/dfu_write_application.py

Challenges:

* It needs to be checked that logging works for the bootloader.
* It needs to be checked at which step the DFU process fails.