Microapp failures properly detected
+ Watchdog stores data in particular RAM section
+ This data is written before and after each microapp call
+ It will be persisted across warm boots.
+ After a boot it is detectable if we were within a microapp when the
  reboot happens.

On the moment it is a one-strike out system. Even if something else
triggered the watchdog (e.g. a while-loop within an interrupt in
bluenet) the microapp will be considered guilty. This is also the case
for a physical reboot.

This might be improved by being slightly less strict. For example, a
reboot while in microapp context increments a counter. Then, when
another reboot happens while not in microapp context this counter is
reset. Only when two successive reboots happen in a microapp context it
is considered guilty. Note, that also such a method has a disadvantage.
If a microapp is faulty within a soft interrupt there might be
considerable time between reboots. Or if there is an error on parsing
particular data it might happen only so now and then. Hence, for now
there's chosen to be strict. The user should just re-enable the microapp.