Cope with more wide range of faults in microapp
Currently we do not just use the watchdog timer, but we set each time
the operating state in IPC ram. This means that even if the watchdog
timer is NOT triggered, we can still detect that a microapp has been
at fault (likely).

This means that the Watchdog class can best be renamed to something more
general such as RebootManager. It is responsible for information that
has to be preserved across warm reboots. The use of the watchdog is
secondary.

The callback handler for the watchdog is now empty. It is not removed
because it might be worth looking into actually also recognizing
watchdog events and storing those as well (apart from microapps).