Option to build offline
Solves https://github.com/crownstone/bluenet/issues/176

Through the option `-DBUILD_OFFLINE=ON` you can disable updates of for
example the git repositories added by ExternalProject_Add.

Of course it's the responsibility of the developer to have the required
dependencies such as the cross-compiler on their system. This option is
only useful after you have compiled before with this option disabled
(which pulls the corresponding dependencies into the project).