Fix bugs revealed by reordering by clang-format
The macro APP_ERROR_HANDLER was defined or not defined depending on the
order of header inclusion for the test files.

Now we just implemented the stubs app_error_handler and
app_error_handler_bare so that they can be used on the laptop. The
error_code argument is used as the exit state of the process.