Warn when baudrate is too high for given JLink firmware version
Generate a CMake warning when a baudrate of more than 115200 is used and
there is a JLink present which has firmware that is too new. From V7.00
onwards there is an artificial limitation to 115200 baud because a JLink
might accidentally lockup at higher rates.

Crownstone firmware by default uses 230400 as a baudrate. To downgrade
this to something lower means updating python libraries or software
using those libraries (Home Assistant).

Rather it is recommended to flash firmware V6.98e to the JLinks that
are used for testing or programming in the factory.

See: https://forum.segger.com/index.php/Thread/7963-SOLVED-VCOM-does-not-work-after-FW-update-in-7-00/